fix css and move calendar into calendarview
