+++ package.json
@@ -18,7 +18,7 @@
     "react-router": "^4.3.1",
     "react-scripts": "^2.1.5",
     "react-scrollbar": "^0.5.4",
-    "react-select": "^2.3.0",
+    "react-select": "^2.4.1",
     "redux": "^4.0.1"
   },
   "scripts": {
