+++ src/App.css
@@ -12,7 +12,7 @@ import Example.css here and it can be used in Example.js
 @import '../src/components/Footer/Footer.css';
 @import '../src/components/Header/Header.css';
 @import '../src/components/Header/Menu/Menu.css';
-@import '../src/pages/View/View.css'
+@import '../src/pages/View/View.css';
 
 .App {
   text-align: center;
