+++ src/pages/View/CalendarView/CalendarView.js
@@ -1,4 +1,8 @@
 import React, {Component} from 'react';
+import moment from 'moment';
+import BigCalendar from 'react-big-calendar';
+
+const localizer = BigCalendar.momentLocalizer(moment)
 
 export default class CalendarView extends Component {
     constructor(props) {
@@ -7,8 +11,24 @@ export default class CalendarView extends Component {
     }
 
     render() {
+        let myEventsList = [];
+        let event1 = {
+            title: "Testing",
+            start: new Date(),
+            end: new Date(),
+            "allDay?": true,
+            "resource?": "Hello"
+        }
+        myEventsList.push(event1);
+        
         return (
-            <div>calendar view</div>
+            <div className="box">
+                <BigCalendar
+                    localizer={localizer}
+                    events={myEventsList}
+                    startAccessor="start"
+                    endAccessor="end"/>
+            </div>
         );
     }
 }
