+++ src/pages/View/View.js
@@ -1,9 +1,9 @@
 import React, {Component} from 'react';
 import ChartView from './ChartView/ChartView';
 import CalendarView from './CalendarView/CalendarView';
-import './View.css';
 import moment from 'moment';
-import BigCalendar from 'react-big-calendar'
+import BigCalendar from 'react-big-calendar';
+import './View.css';
 
 const localizer = BigCalendar.momentLocalizer(moment)
 
@@ -20,21 +20,10 @@ export default class View extends Component {
     }
 
     render() {
-
-        let myEventsList = [];
-        let event1 = {
-            title: "Testing",
-            start: new Date(),
-            end: new Date(),
-            "allDay?": true,
-            "resource?": "Hello"
-        }
-        myEventsList.push(event1);
-
         return (
             <div className="view">
                 <div>
-                    <h2>Available Times</h2>
+                    <h2 className="heading2">Available Times</h2>
                     <button
                         onClick={this.toggleView}
                         data={this.props.data}
@@ -54,13 +43,6 @@ export default class View extends Component {
                             ? <ChartView/>
                             : <CalendarView/>
                     }
-                    {/* <div className="box">
-                        <BigCalendar
-                            localizer={localizer}
-                            events={myEventsList}
-                            startAccessor="start"
-                            endAccessor="end"/>
-                    </div> */}
                 </div>
             </div>
         );
